% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.conditions.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design is valid}
\usage{
check.conditions(conds)
}
\arguments{
\item{conds}{A vector}
}
\value{
A list
}
\description{
This function check the validity of the conditions
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata')
check.conditions(Biobase::pData(Exp1_R25_pept)$Condition)
}

}
\author{
Samuel Wieczorek originally in the DAPAR package. Included in this 
package since DAPAR is to be removed from Bioconductor >= 3.15.
}
