% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasim}
\alias{datasim}
\title{A single simulated dataset}
\format{
A data frame with 200 observations on the following 11 variables.
\describe{ 
\item{id.obs}{a numeric vector} 
\item{X1}{a numeric vector} 
\item{X2}{a numeric vector} 
\item{X3}{a numeric vector} 
\item{X4}{a numeric vector} 
\item{X5}{a numeric vector} 
\item{X6}{a numeric vector} 
\item{X7}{a numeric vector} 
\item{X8}{a numeric vector} 
\item{X9}{a numeric vector} 
\item{X10}{a numeric vector} 
}
}
\source{
We simulated the data.
}
\description{
This dataset was simulated using the default values of the 
values of the options of the protdatasim function and the set.seed value set 
to 4619.
}
\examples{

data(datasim)
str(datasim)

}
\references{
M. Chion, Ch. Carapito and F. Bertrand (2021). \emph{Accounting for multiple imputation-induced variability for differential analysis in mass spectrometry-based label-free quantitative proteomics}. \doi{doi:10.1371/journal.pcbi.1010420}.
}
\author{
M. Chion, Ch. Carapito and F. Bertrand.
}
\keyword{datasets}
