% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBayes_mod.R
\name{eBayes.mod}
\alias{eBayes.mod}
\title{Title}
\usage{
eBayes.mod(
  fit,
  VarRubin,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1)
)
}
\arguments{
\item{fit}{}

\item{VarRubin}{}

\item{proportion}{}

\item{stdev.coef.lim}{}

\item{trend}{}

\item{robust}{}

\item{winsor.tail.p}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
