% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaCompleteTest_mod.R
\name{limmaCompleteTest.mod}
\alias{limmaCompleteTest.mod}
\title{Title}
\usage{
limmaCompleteTest.mod(qData, sTab, comp.type = "OnevsOne")
}
\arguments{
\item{qData}{}

\item{sTab}{}

\item{comp.type}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
