% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.design.R
\name{make.design.3}
\alias{make.design.3}
\title{Builds the design matrix for designs of level 3}
\usage{
make.design.3(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the pData function of MSnbase}
}
\value{
A design matrix
}
\description{
This function builds the design matrix for design of level 3
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata')
sTab <-cbind(Biobase::pData(Exp1_R25_pept), Tech.Rep=1:6)
make.design.3(sTab)
}

}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek originally in 
the DAPAR package. Included in this package since DAPAR was to be removed from 
Bioconductor >= 3.15.
}
