% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanImp_emmeans.R
\name{meanImp_emmeans}
\alias{meanImp_emmeans}
\title{Title}
\usage{
meanImp_emmeans(ind, peptide = 1, tabdata, metacond)
}
\arguments{
\item{ind}{}

\item{peptide}{}

\item{tabdata}{}

\item{metacond}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
