% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi4limma.R
\name{mi4limma}
\alias{mi4limma}
\title{Title}
\usage{
mi4limma(qData, sTab, VarRubin, comp.type = "OnevsOne", robust = F)
}
\arguments{
\item{qData}{}

\item{sTab}{}

\item{VarRubin}{}

\item{comp.type}{}

\item{robust}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
