% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi4p-package.R
\docType{package}
\name{mi4p-package}
\alias{mi4p-package}
\alias{mi4p}
\title{The mi4p Package}
\description{
A multiple imputation framework ...
}
\references{
M. Chion, Ch. Carapito, F. Bertrand, arxiv preprint.
}
\author{
This package has been written by Marie Chion, Christine Carapito and Frederic Bertrand. 
Maintainer: <frederic.bertrand@utt.fr>
}
\keyword{package}
