% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mm_peptides}
\alias{mm_peptides}
\title{mm_peptides - peptide-level intensities for mouse}
\format{
A data frame with 1102 rows and 13 colummns, compiring 7 columns of
        metadata and 6 columns of peptide intensities. 69 proteins.
}
\usage{
data(mm_peptides)
}
\description{
A dataset containing the protein and petide information and peptide-level
intensities for 6 samples: 3 CG and 3 mCG groups. There are 69 proteins.
The columns are as follows:
}
\details{
\itemize{
  \item Sequence - peptide sequence - randomly chosen from a larger list of
        sequences
  \item MatchedID - numeric ID that links proteins in the two datasets,
        unnecessary if datasets are for the same species
  \item ProtID - protein ID, artificial protein ID, eg. Prot1, Prot2, ...
  \item GeneID - gene ID, artificial gene ID, eg. Gene1, Gene2, ...
  \item ProtName - artificial Protein Name
  \item ProtIDLong - long protein ID, full protein name, here artificially
        simulated
  \item GeneIDLong - long gene ID, full gene name, here artificially
        simulated
  \item CG1 - raw intensity column for sample 1 in CG group
  \item CG2 - raw intensity column for sample 2 in CG group
  \item CG3 - raw intensity column for sample 3 in CG group
  \item mCG1 - raw intensity column for sample 1 in mCG group
  \item mCG2 - raw intensity column for sample 2 in mCG group
  \item mCG3 - raw intensity column for sample 3 in mCG group
}
}
\keyword{datasets}
