% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{norm.200.m100.sd1.vs.m200.sd1.list}
\alias{norm.200.m100.sd1.vs.m200.sd1.list}
\title{A list of simulated datasets.}
\format{
The format is: List of 100 $ :'data.frame': 200 obs. of 11
variables: ..$ id.obs: int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num
[1:200] 99.6 99.9 100.2 99.8 100.4 ...  ..$ X2 : num [1:200] 97.4 101.3
100.3 100.2 101.7 ...  ..$ X3 : num [1:200] 100.3 100.9 99.1 101.2 100.6 ...
..$ X4 : num [1:200] 99.4 99.2 98.5 99.1 99.5 ...  ..$ X5 : num [1:200] 98.5
99.7 100 100.2 100.7 ...  ..$ X6 : num [1:200] 200 199 199 200 199 ...  ..$
X7 : num [1:200] 200 200 202 199 199 ...  ..$ X8 : num [1:200] 202 199 200
199 201 ...  ..$ X9 : num [1:200] 200 200 199 201 200 ...  ..$ X10 : num
[1:200] 200 198 200 201 199 ...  ..- attr(*, "metadata")='data.frame': 10
obs. of 3 variables: .. ..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...
.. ..$ Condition : Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$
Bio.Rep : int [1:10] 1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11
variables: ..$ id.obs: int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num
[1:200] 100.3 97.6 100.1 99.3 101.1 ...  ..$ X2 : num [1:200] 100.4 98.6
99.9 99.7 100.2 ...  ..$ X3 : num [1:200] 100.4 101.3 99.1 100.2 100.3 ...
..$ X4 : num [1:200] 98.9 100.5 101 100.5 99.5 ...  ..$ X5 : num [1:200]
101.5 99.7 98.1 99.5 100.2 ...  ..$ X6 : num [1:200] 201 200 201 201 200 ...
..$ X7 : num [1:200] 200 200 201 200 199 ...  ..$ X8 : num [1:200] 198 200
202 199 199 ...  ..$ X9 : num [1:200] 200 201 198 200 200 ...  ..$ X10 : num
[1:200] 200 199 201 199 199 ...  ..- attr(*, "metadata")='data.frame': 10
obs. of 3 variables: .. ..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...
.. ..$ Condition : Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$
Bio.Rep : int [1:10] 1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11
variables: ..$ id.obs: int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num
[1:200] 101 98.8 101 101.2 100.1 ...  ..$ X2 : num [1:200] 101.1 100.6 98.8
100.8 99.9 ...  ..$ X3 : num [1:200] 99.6 99.7 100.1 98.7 101.4 ...  ..$ X4
: num [1:200] 100.3 100.6 100.1 99.5 97.4 ...  ..$ X5 : num [1:200] 100.4
100.1 101 99.1 100.3 ...  ..$ X6 : num [1:200] 202 200 200 200 202 ...  ..$
X7 : num [1:200] 198 199 201 200 199 ...  ..$ X8 : num [1:200] 202 201 200
201 198 ...  ..$ X9 : num [1:200] 199 200 200 199 199 ...  ..$ X10 : num
[1:200] 199 199 201 201 199 ...  ..- attr(*, "metadata")='data.frame': 10
obs. of 3 variables: .. ..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...
.. ..$ Condition : Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$
Bio.Rep : int [1:10] 1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11
variables: ..$ id.obs: int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num
[1:200] 98.4 98.5 100.5 98.7 101.4 ...  ..$ X2 : num [1:200] 98.4 101.2 99.1
100.4 99.3 ...  ..$ X3 : num [1:200] 99.8 100.3 98.4 99.6 101 ...  ..$ X4 :
num [1:200] 98.5 98.8 99.8 100.4 98.7 ...  ..$ X5 : num [1:200] 100.9 99.5
100.5 100.5 99.3 ...  ..$ X6 : num [1:200] 201 200 200 198 199 ...  ..$ X7 :
num [1:200] 201 200 201 202 200 ...  ..$ X8 : num [1:200] 201 200 200 201
199 ...  ..$ X9 : num [1:200] 201 198 201 201 198 ...  ..$ X10 : num [1:200]
201 200 199 200 200 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3
variables: .. ..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$
Condition : Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep :
int [1:10] 1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables:
..$ id.obs: int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.2
100.7 99.6 99.8 100.1 ...  ..$ X2 : num [1:200] 100.8 98.7 101.1 101.8 99.9
...  ..$ X3 : num [1:200] 101.5 100.6 100.9 100.3 99.8 ...  ..$ X4 : num
[1:200] 97.3 99.1 100.2 98.4 98.4 ...  ..$ X5 : num [1:200] 101 101 101 100
101 ...  ..$ X6 : num [1:200] 201 201 198 200 199 ...  ..$ X7 : num [1:200]
198 199 199 201 200 ...  ..$ X8 : num [1:200] 199 197 199 199 199 ...  ..$
X9 : num [1:200] 201 199 199 199 200 ...  ..$ X10 : num [1:200] 201 199 199
200 197 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables:
.. ..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.6 100.7 100.1
99 101 ...  ..$ X2 : num [1:200] 99.1 99.9 100.9 99.9 99.8 ...  ..$ X3 : num
[1:200] 99.4 100.3 100.6 100.7 98.6 ...  ..$ X4 : num [1:200] 100 100.3
100.2 98.2 103.7 ...  ..$ X5 : num [1:200] 100.7 99.1 100.8 100.4 99.5 ...
..$ X6 : num [1:200] 201 201 199 199 201 ...  ..$ X7 : num [1:200] 201 199
202 202 202 ...  ..$ X8 : num [1:200] 201 199 199 202 201 ...  ..$ X9 : num
[1:200] 200 199 200 202 200 ...  ..$ X10 : num [1:200] 200 201 200 200 203
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.9 98.9 99.4 98.4
98.6 ...  ..$ X2 : num [1:200] 100.3 99.4 101.3 99.4 100.2 ...  ..$ X3 : num
[1:200] 99 101 102 101 102 ...  ..$ X4 : num [1:200] 99.5 100.2 99.4 100.5
99.7 ...  ..$ X5 : num [1:200] 98.8 100.4 100.2 100 99.9 ...  ..$ X6 : num
[1:200] 200 200 201 200 202 ...  ..$ X7 : num [1:200] 199 202 201 199 200
...  ..$ X8 : num [1:200] 201 201 200 200 200 ...  ..$ X9 : num [1:200] 201
200 200 201 200 ...  ..$ X10 : num [1:200] 202 199 201 201 202 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.7 99.1 99.1 98.3
98.9 ...  ..$ X2 : num [1:200] 100.7 100.1 98.4 100.9 100.2 ...  ..$ X3 :
num [1:200] 101.1 101 100.9 99.2 101.5 ...  ..$ X4 : num [1:200] 99.7 99
100.2 101.2 98.1 ...  ..$ X5 : num [1:200] 100.3 101.3 100.1 99.6 100.1 ...
..$ X6 : num [1:200] 200 200 197 200 200 ...  ..$ X7 : num [1:200] 199 198
199 200 199 ...  ..$ X8 : num [1:200] 200 201 201 198 199 ...  ..$ X9 : num
[1:200] 200 201 200 200 201 ...  ..$ X10 : num [1:200] 199 203 203 199 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.8 98.9 98.8 99.2
99.5 ...  ..$ X2 : num [1:200] 101.6 98.7 100 101 100 ...  ..$ X3 : num
[1:200] 98.5 99.6 100.6 101.3 100 ...  ..$ X4 : num [1:200] 100.3 100.5 101
99.1 99.1 ...  ..$ X5 : num [1:200] 100.2 98.2 100.3 99.1 99.6 ...  ..$ X6 :
num [1:200] 201 201 200 199 199 ...  ..$ X7 : num [1:200] 200 199 200 198
198 ...  ..$ X8 : num [1:200] 200 201 200 199 200 ...  ..$ X9 : num [1:200]
201 200 201 200 201 ...  ..$ X10 : num [1:200] 201 200 200 201 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.4 99.9 99.7 99.1
100.8 ...  ..$ X2 : num [1:200] 101 101 100 102 99 ...  ..$ X3 : num [1:200]
100.8 99.3 100.7 99.9 100 ...  ..$ X4 : num [1:200] 98.9 99.8 100.4 100.6
101.3 ...  ..$ X5 : num [1:200] 100.5 99.9 100.8 99 99.5 ...  ..$ X6 : num
[1:200] 201 198 199 199 201 ...  ..$ X7 : num [1:200] 199 200 200 201 200
...  ..$ X8 : num [1:200] 201 198 200 202 202 ...  ..$ X9 : num [1:200] 199
200 200 200 201 ...  ..$ X10 : num [1:200] 200 199 200 200 202 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.4 99.6 99.8 100.6
99.8 ...  ..$ X2 : num [1:200] 97 101 101 100 101 ...  ..$ X3 : num [1:200]
99.6 98.6 100 102 101 ...  ..$ X4 : num [1:200] 99.2 99.1 100.6 99.8 98.7
...  ..$ X5 : num [1:200] 100.3 101.4 99.3 99.4 98.7 ...  ..$ X6 : num
[1:200] 200 198 198 199 201 ...  ..$ X7 : num [1:200] 199 199 199 201 199
...  ..$ X8 : num [1:200] 200 200 201 201 201 ...  ..$ X9 : num [1:200] 200
200 199 200 198 ...  ..$ X10 : num [1:200] 201 200 201 199 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.1 98 100.4 98.9
97.8 ...  ..$ X2 : num [1:200] 98 99.1 100.5 100.3 99.3 ...  ..$ X3 : num
[1:200] 98.9 98.8 99.9 101.8 99.4 ...  ..$ X4 : num [1:200] 101 100 101 102
100 ...  ..$ X5 : num [1:200] 100.7 101.9 100.7 97.6 99.3 ...  ..$ X6 : num
[1:200] 198 199 199 200 202 ...  ..$ X7 : num [1:200] 198 200 199 202 200
...  ..$ X8 : num [1:200] 201 202 200 200 200 ...  ..$ X9 : num [1:200] 199
200 198 200 198 ...  ..$ X10 : num [1:200] 199 199 198 199 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.6 100 101.4 99.3
101.9 ...  ..$ X2 : num [1:200] 101.4 98.1 101.2 101 99.5 ...  ..$ X3 : num
[1:200] 99 100.3 100.5 102.7 98.5 ...  ..$ X4 : num [1:200] 99.5 99.3 99.4
100.7 100 ...  ..$ X5 : num [1:200] 99.7 99.9 100 100.2 98.6 ...  ..$ X6 :
num [1:200] 199 201 198 201 199 ...  ..$ X7 : num [1:200] 203 198 200 200
199 ...  ..$ X8 : num [1:200] 200 201 199 201 201 ...  ..$ X9 : num [1:200]
200 200 201 202 200 ...  ..$ X10 : num [1:200] 200 198 201 201 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.8 100.4 100.5
99.5 100.8 ...  ..$ X2 : num [1:200] 99.9 100.3 101.5 100.2 99.9 ...  ..$ X3
: num [1:200] 100.5 98.9 99.5 100.3 101.1 ...  ..$ X4 : num [1:200] 99.5
100.5 99.4 101.6 101 ...  ..$ X5 : num [1:200] 100.9 100.1 99.2 99.2 99.1
...  ..$ X6 : num [1:200] 202 200 199 201 200 ...  ..$ X7 : num [1:200] 199
201 200 198 201 ...  ..$ X8 : num [1:200] 198 200 199 200 201 ...  ..$ X9 :
num [1:200] 202 199 202 200 200 ...  ..$ X10 : num [1:200] 199 199 202 198
201 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.7 98.6 99.6
101.2 98.5 ...  ..$ X2 : num [1:200] 99.9 97.7 100.1 100 99.2 ...  ..$ X3 :
num [1:200] 100 100 101 101 100 ...  ..$ X4 : num [1:200] 100 101.5 99.9
100.9 99 ...  ..$ X5 : num [1:200] 99.2 100.4 97.5 99.8 99.6 ...  ..$ X6 :
num [1:200] 199 199 199 200 199 ...  ..$ X7 : num [1:200] 199 199 201 201
199 ...  ..$ X8 : num [1:200] 201 199 201 200 200 ...  ..$ X9 : num [1:200]
199 201 201 199 199 ...  ..$ X10 : num [1:200] 201 200 201 199 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.3 99 99.8 102.2
100.5 ...  ..$ X2 : num [1:200] 101.8 99.8 98.8 100.5 98.8 ...  ..$ X3 : num
[1:200] 100 101.6 100.9 99.7 98.2 ...  ..$ X4 : num [1:200] 100 100.6 100.4
99.9 99.8 ...  ..$ X5 : num [1:200] 99.5 100.8 99.3 101.6 99.3 ...  ..$ X6 :
num [1:200] 200 200 200 201 199 ...  ..$ X7 : num [1:200] 200 200 201 200
199 ...  ..$ X8 : num [1:200] 200 199 202 199 199 ...  ..$ X9 : num [1:200]
201 201 200 200 200 ...  ..$ X10 : num [1:200] 198 199 199 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.8 99.4 100.1 99.9
100.6 ...  ..$ X2 : num [1:200] 100.2 98.2 99.7 99.8 100.8 ...  ..$ X3 : num
[1:200] 100 100 101 102 101 ...  ..$ X4 : num [1:200] 99.7 99.2 99.9 100.3
100.3 ...  ..$ X5 : num [1:200] 100.1 100.7 100 99.1 100.5 ...  ..$ X6 : num
[1:200] 201 201 198 199 201 ...  ..$ X7 : num [1:200] 202 198 200 199 199
...  ..$ X8 : num [1:200] 200 200 199 201 200 ...  ..$ X9 : num [1:200] 200
202 200 200 201 ...  ..$ X10 : num [1:200] 200 201 200 200 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.3 99.6 100.6
100.6 99.4 ...  ..$ X2 : num [1:200] 99.9 101.3 99.4 100.4 99.5 ...  ..$ X3
: num [1:200] 100.1 99.4 100.5 100.5 101 ...  ..$ X4 : num [1:200] 99.7
100.9 99.4 100.1 100 ...  ..$ X5 : num [1:200] 99.5 100.2 100.9 101.3 100.1
...  ..$ X6 : num [1:200] 199 200 200 201 200 ...  ..$ X7 : num [1:200] 200
198 199 199 201 ...  ..$ X8 : num [1:200] 201 200 201 199 198 ...  ..$ X9 :
num [1:200] 202 200 200 200 200 ...  ..$ X10 : num [1:200] 200 199 201 201
200 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.4 100.2 100.5
99.3 98.9 ...  ..$ X2 : num [1:200] 99.7 100.4 101 101.4 101 ...  ..$ X3 :
num [1:200] 99.8 101.1 99.2 98.8 100.5 ...  ..$ X4 : num [1:200] 99.2 99.5
98.7 99 100.5 ...  ..$ X5 : num [1:200] 100.4 99.5 100.2 98.8 99.1 ...  ..$
X6 : num [1:200] 199 198 198 199 200 ...  ..$ X7 : num [1:200] 200 199 199
199 201 ...  ..$ X8 : num [1:200] 201 199 200 202 201 ...  ..$ X9 : num
[1:200] 199 200 200 198 200 ...  ..$ X10 : num [1:200] 201 200 199 200 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101 99.4 100.3 99.4
100.6 ...  ..$ X2 : num [1:200] 98.6 98.5 100.7 100.5 99.2 ...  ..$ X3 : num
[1:200] 99.4 98.2 99.8 100 101.8 ...  ..$ X4 : num [1:200] 102 101 101 102
101 ...  ..$ X5 : num [1:200] 100.7 101.6 99 99.2 98.8 ...  ..$ X6 : num
[1:200] 199 201 200 200 202 ...  ..$ X7 : num [1:200] 200 200 199 198 199
...  ..$ X8 : num [1:200] 201 198 200 201 200 ...  ..$ X9 : num [1:200] 200
200 200 200 199 ...  ..$ X10 : num [1:200] 200 200 201 200 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.4 98.5 100.5 99.9
100.7 ...  ..$ X2 : num [1:200] 100.4 100.3 98.7 100 99.6 ...  ..$ X3 : num
[1:200] 99.3 100 99.3 99.2 99.7 ...  ..$ X4 : num [1:200] 98.6 100.3 100.4
100.7 99.4 ...  ..$ X5 : num [1:200] 100.3 99.9 101.9 100.5 101.3 ...  ..$
X6 : num [1:200] 200 199 200 199 200 ...  ..$ X7 : num [1:200] 200 202 199
202 199 ...  ..$ X8 : num [1:200] 199 201 200 200 200 ...  ..$ X9 : num
[1:200] 200 200 202 200 201 ...  ..$ X10 : num [1:200] 201 201 201 201 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.7 99.9 101.9 99.4
100.7 ...  ..$ X2 : num [1:200] 102.3 100.3 100.2 97.7 100 ...  ..$ X3 : num
[1:200] 99.7 98.8 100.1 100.1 100.7 ...  ..$ X4 : num [1:200] 100 100.3
100.7 100.8 98.9 ...  ..$ X5 : num [1:200] 100.2 100.3 99.7 99.5 100.6 ...
..$ X6 : num [1:200] 199 200 199 200 201 ...  ..$ X7 : num [1:200] 199 198
200 200 201 ...  ..$ X8 : num [1:200] 200 198 201 200 202 ...  ..$ X9 : num
[1:200] 201 200 201 200 199 ...  ..$ X10 : num [1:200] 201 198 200 200 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.7 100.7 99.3 98
100.2 ...  ..$ X2 : num [1:200] 99.2 100.8 99.9 100.3 99.7 ...  ..$ X3 : num
[1:200] 100.4 102.2 99.5 100.1 100.4 ...  ..$ X4 : num [1:200] 97.4 101.6
99.5 100.8 99.7 ...  ..$ X5 : num [1:200] 100.3 100.7 101.4 99.8 100.8 ...
..$ X6 : num [1:200] 200 199 201 198 200 ...  ..$ X7 : num [1:200] 201 201
202 201 199 ...  ..$ X8 : num [1:200] 201 202 200 203 201 ...  ..$ X9 : num
[1:200] 200 200 200 200 201 ...  ..$ X10 : num [1:200] 201 200 201 200 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.6 100.8 101.2
100.6 99.6 ...  ..$ X2 : num [1:200] 100.8 99.8 101.4 100.7 98.9 ...  ..$ X3
: num [1:200] 99.9 100.9 99.1 99.8 100.9 ...  ..$ X4 : num [1:200] 98.9
100.3 98.7 98.8 97.9 ...  ..$ X5 : num [1:200] 101.2 100.5 100.7 98.6 100.1
...  ..$ X6 : num [1:200] 200 200 201 200 200 ...  ..$ X7 : num [1:200] 201
199 202 200 200 ...  ..$ X8 : num [1:200] 200 199 200 201 199 ...  ..$ X9 :
num [1:200] 200 201 200 200 203 ...  ..$ X10 : num [1:200] 200 201 200 199
200 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100 99.9 100.1
101.8 100.9 ...  ..$ X2 : num [1:200] 99.9 100.9 98.7 100.6 100.2 ...  ..$
X3 : num [1:200] 97.4 99.9 99.7 98.2 99.9 ...  ..$ X4 : num [1:200] 101 98.5
99.8 100.4 101.5 ...  ..$ X5 : num [1:200] 101.3 99.7 99.2 101.2 99.5 ...
..$ X6 : num [1:200] 199 199 199 200 200 ...  ..$ X7 : num [1:200] 199 200
202 199 198 ...  ..$ X8 : num [1:200] 199 200 201 199 197 ...  ..$ X9 : num
[1:200] 201 200 198 200 199 ...  ..$ X10 : num [1:200] 200 199 200 201 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.5 99.7 99.7 99.9
100.9 ...  ..$ X2 : num [1:200] 100.3 100.8 99.2 99.1 100 ...  ..$ X3 : num
[1:200] 100.8 98.8 98.7 100.5 99.6 ...  ..$ X4 : num [1:200] 101.1 100.1
100.7 99.9 100.5 ...  ..$ X5 : num [1:200] 97.4 100.3 100.9 101.6 99.3 ...
..$ X6 : num [1:200] 199 201 198 199 199 ...  ..$ X7 : num [1:200] 199 201
200 201 200 ...  ..$ X8 : num [1:200] 201 201 199 200 198 ...  ..$ X9 : num
[1:200] 202 199 199 201 202 ...  ..$ X10 : num [1:200] 200 203 201 199 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.6 99.8 98.7 98.5
99.7 ...  ..$ X2 : num [1:200] 102 97.7 99.6 99.2 96.4 ...  ..$ X3 : num
[1:200] 99.9 101.3 100.1 99.6 99.1 ...  ..$ X4 : num [1:200] 98.9 99 99.1
98.6 99.2 ...  ..$ X5 : num [1:200] 99.3 99.2 100.2 100.4 100.3 ...  ..$ X6
: num [1:200] 202 200 201 201 201 ...  ..$ X7 : num [1:200] 201 200 200 199
199 ...  ..$ X8 : num [1:200] 200 200 200 200 199 ...  ..$ X9 : num [1:200]
201 200 199 200 202 ...  ..$ X10 : num [1:200] 201 201 200 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.7 101.3 99.8 100.2
100.2 ...  ..$ X2 : num [1:200] 100.3 98.3 99.2 100.6 100.3 ...  ..$ X3 :
num [1:200] 98.5 97.9 101 98.2 98.7 ...  ..$ X4 : num [1:200] 100.2 98.8
100.8 99.6 100.5 ...  ..$ X5 : num [1:200] 99.1 98.7 99.5 100.8 99.4 ...
..$ X6 : num [1:200] 201 201 199 201 199 ...  ..$ X7 : num [1:200] 201 199
202 200 200 ...  ..$ X8 : num [1:200] 200 199 202 200 199 ...  ..$ X9 : num
[1:200] 200 200 202 201 200 ...  ..$ X10 : num [1:200] 199 200 201 200 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.1 99.6 100.4 101.8
99.3 ...  ..$ X2 : num [1:200] 98.9 99.5 100.7 100.4 100.2 ...  ..$ X3 : num
[1:200] 101.9 99.8 98.6 99.4 101.3 ...  ..$ X4 : num [1:200] 100.8 101.9
99.9 101.4 100.2 ...  ..$ X5 : num [1:200] 101.6 99.2 99.6 98.9 101.8 ...
..$ X6 : num [1:200] 200 200 200 200 200 ...  ..$ X7 : num [1:200] 201 199
200 201 201 ...  ..$ X8 : num [1:200] 200 201 199 200 201 ...  ..$ X9 : num
[1:200] 201 200 200 200 201 ...  ..$ X10 : num [1:200] 201 200 202 200 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 102 100 101 100 101
...  ..$ X2 : num [1:200] 100.3 100.8 99.7 99.6 101 ...  ..$ X3 : num
[1:200] 99.7 99.6 101.4 101.9 101 ...  ..$ X4 : num [1:200] 100 99.6 99.2
98.6 100.1 ...  ..$ X5 : num [1:200] 98.9 100.3 99.8 100.7 98.8 ...  ..$ X6
: num [1:200] 200 199 199 201 200 ...  ..$ X7 : num [1:200] 199 201 201 199
201 ...  ..$ X8 : num [1:200] 199 201 201 201 201 ...  ..$ X9 : num [1:200]
200 200 198 199 200 ...  ..$ X10 : num [1:200] 199 199 201 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.6 99.2 99.8 99.9
98.6 ...  ..$ X2 : num [1:200] 99.6 100.5 100.6 99.6 100.5 ...  ..$ X3 : num
[1:200] 100.6 100.1 99.5 100.8 99.8 ...  ..$ X4 : num [1:200] 101.2 100.1
100.7 99.1 100.4 ...  ..$ X5 : num [1:200] 101.4 100.2 100.2 99.9 100.1 ...
..$ X6 : num [1:200] 200 199 200 200 200 ...  ..$ X7 : num [1:200] 199 199
201 200 201 ...  ..$ X8 : num [1:200] 200 200 200 201 199 ...  ..$ X9 : num
[1:200] 200 200 201 199 200 ...  ..$ X10 : num [1:200] 201 200 201 200 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.7 101.3 99.1 99.3
98.8 ...  ..$ X2 : num [1:200] 101.8 100.8 99.2 100.5 98.9 ...  ..$ X3 : num
[1:200] 99.2 98.1 100.4 100.2 97.9 ...  ..$ X4 : num [1:200] 99.5 99.8 99.8
99.8 99.2 ...  ..$ X5 : num [1:200] 99.9 99.8 98.6 98.7 99.1 ...  ..$ X6 :
num [1:200] 199 198 200 199 200 ...  ..$ X7 : num [1:200] 199 200 201 200
199 ...  ..$ X8 : num [1:200] 199 200 200 200 200 ...  ..$ X9 : num [1:200]
200 200 202 199 199 ...  ..$ X10 : num [1:200] 199 198 199 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.8 101.2 100.5 98.1
99.2 ...  ..$ X2 : num [1:200] 100.1 99.7 99.7 100 100.6 ...  ..$ X3 : num
[1:200] 100.7 98.4 100.3 100.1 99.8 ...  ..$ X4 : num [1:200] 100.3 99.8 99
99 100 ...  ..$ X5 : num [1:200] 99.8 100.8 101.3 100.4 101.6 ...  ..$ X6 :
num [1:200] 200 200 200 199 200 ...  ..$ X7 : num [1:200] 201 199 201 199
200 ...  ..$ X8 : num [1:200] 200 198 199 201 199 ...  ..$ X9 : num [1:200]
200 200 200 200 200 ...  ..$ X10 : num [1:200] 201 201 199 200 202 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.4 100.6 99.9 98.2
99.9 ...  ..$ X2 : num [1:200] 98.6 97.6 101 100 100.6 ...  ..$ X3 : num
[1:200] 99.7 98.1 100.9 100.4 99 ...  ..$ X4 : num [1:200] 99.1 99 99.7 99.8
100.2 ...  ..$ X5 : num [1:200] 99 101.2 99.5 99.1 100.9 ...  ..$ X6 : num
[1:200] 199 198 201 200 201 ...  ..$ X7 : num [1:200] 200 199 200 201 201
...  ..$ X8 : num [1:200] 200 199 200 201 202 ...  ..$ X9 : num [1:200] 200
200 202 201 201 ...  ..$ X10 : num [1:200] 199 200 199 200 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.5 98.6 100 101
98.9 ...  ..$ X2 : num [1:200] 99.3 99.5 100.9 99.7 99.2 ...  ..$ X3 : num
[1:200] 101.7 100.1 97.5 99.1 99.5 ...  ..$ X4 : num [1:200] 100.5 99.6 99.1
100.3 100.2 ...  ..$ X5 : num [1:200] 99.7 99.2 98.9 100.1 99.4 ...  ..$ X6
: num [1:200] 199 200 201 200 199 ...  ..$ X7 : num [1:200] 201 199 199 201
200 ...  ..$ X8 : num [1:200] 199 199 199 200 201 ...  ..$ X9 : num [1:200]
201 199 202 199 199 ...  ..$ X10 : num [1:200] 199 200 199 199 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 97.8 100.8 101.8
100.3 ...  ..$ X2 : num [1:200] 101.3 99.6 98.8 99.7 99.5 ...  ..$ X3 : num
[1:200] 100.8 98.1 99.8 101.3 99.5 ...  ..$ X4 : num [1:200] 100.3 99.4 100
99.2 100.5 ...  ..$ X5 : num [1:200] 100.8 100.1 100.3 101.1 99.4 ...  ..$
X6 : num [1:200] 200 199 201 201 200 ...  ..$ X7 : num [1:200] 201 201 200
200 200 ...  ..$ X8 : num [1:200] 201 199 199 200 201 ...  ..$ X9 : num
[1:200] 199 200 198 200 199 ...  ..$ X10 : num [1:200] 199 201 199 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.2 99.7 100.5
100.1 101.2 ...  ..$ X2 : num [1:200] 99.9 101.2 99.2 100.5 100 ...  ..$ X3
: num [1:200] 101.7 99.8 100.1 99.6 98.9 ...  ..$ X4 : num [1:200] 98.4 98.3
101.3 100.5 100.1 ...  ..$ X5 : num [1:200] 98.4 98.6 100.5 100 99.7 ...
..$ X6 : num [1:200] 198 200 200 200 200 ...  ..$ X7 : num [1:200] 200 200
199 199 200 ...  ..$ X8 : num [1:200] 200 199 200 200 201 ...  ..$ X9 : num
[1:200] 201 199 201 201 198 ...  ..$ X10 : num [1:200] 199 200 200 200 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.7 100.8 99.3 98.1
98.4 ...  ..$ X2 : num [1:200] 98 100.3 100.1 100.8 99.1 ...  ..$ X3 : num
[1:200] 101 100.4 101.1 100.1 99.2 ...  ..$ X4 : num [1:200] 101 100 99.5
100 99.6 ...  ..$ X5 : num [1:200] 98.4 100.5 99.5 100.6 98.9 ...  ..$ X6 :
num [1:200] 200 199 201 199 199 ...  ..$ X7 : num [1:200] 201 200 201 199
200 ...  ..$ X8 : num [1:200] 200 201 200 200 200 ...  ..$ X9 : num [1:200]
199 200 200 198 199 ...  ..$ X10 : num [1:200] 199 199 199 199 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.4 101 100.6 100.6
98.4 ...  ..$ X2 : num [1:200] 99.4 98.3 99 99.3 99.6 ...  ..$ X3 : num
[1:200] 98.5 100.2 99.7 99.6 99.7 ...  ..$ X4 : num [1:200] 100.1 99.5 100.7
99.1 97.7 ...  ..$ X5 : num [1:200] 98.7 98.5 99 99.5 99.5 ...  ..$ X6 : num
[1:200] 200 200 201 199 199 ...  ..$ X7 : num [1:200] 198 201 202 200 200
...  ..$ X8 : num [1:200] 201 199 200 199 201 ...  ..$ X9 : num [1:200] 201
200 199 200 201 ...  ..$ X10 : num [1:200] 200 201 201 199 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.2 100.9 99.8 99.6
101.2 ...  ..$ X2 : num [1:200] 99.6 100.6 101.6 99.8 100 ...  ..$ X3 : num
[1:200] 101.9 100.1 101.3 99.8 99.2 ...  ..$ X4 : num [1:200] 100.9 100.6
101 99.4 100.3 ...  ..$ X5 : num [1:200] 100.7 100.2 100.8 98.8 100.2 ...
..$ X6 : num [1:200] 200 201 202 200 199 ...  ..$ X7 : num [1:200] 200 201
199 202 201 ...  ..$ X8 : num [1:200] 201 200 201 198 202 ...  ..$ X9 : num
[1:200] 200 200 200 200 201 ...  ..$ X10 : num [1:200] 201 200 200 200 202
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.3 99.4 98.1 99.1
99 ...  ..$ X2 : num [1:200] 98.7 99.6 98 98.3 100.1 ...  ..$ X3 : num
[1:200] 101.8 100.8 98.3 97 98.9 ...  ..$ X4 : num [1:200] 101.3 100.1 101.2
100.1 98.7 ...  ..$ X5 : num [1:200] 99.6 101.3 98.9 99.5 100.9 ...  ..$ X6
: num [1:200] 200 199 200 199 200 ...  ..$ X7 : num [1:200] 202 199 201 201
199 ...  ..$ X8 : num [1:200] 200 201 200 200 200 ...  ..$ X9 : num [1:200]
200 198 201 201 199 ...  ..$ X10 : num [1:200] 201 200 199 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99 99.1 99.9 98.8
100.4 ...  ..$ X2 : num [1:200] 100.4 100 99.5 99.4 100.9 ...  ..$ X3 : num
[1:200] 100 101.4 100.9 99.8 98.3 ...  ..$ X4 : num [1:200] 100.3 98 100.6
100.7 99.7 ...  ..$ X5 : num [1:200] 100 100 100 101 100 ...  ..$ X6 : num
[1:200] 201 199 199 201 199 ...  ..$ X7 : num [1:200] 199 200 199 202 200
...  ..$ X8 : num [1:200] 200 202 201 198 201 ...  ..$ X9 : num [1:200] 199
200 199 198 199 ...  ..$ X10 : num [1:200] 200 201 200 199 198 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.4 100.4 99.8 99.1
99.5 ...  ..$ X2 : num [1:200] 100.2 100.2 99.9 100.9 100.8 ...  ..$ X3 :
num [1:200] 99.6 98.9 99.5 98.6 98.1 ...  ..$ X4 : num [1:200] 100.6 100.3
99.8 99.1 101.3 ...  ..$ X5 : num [1:200] 99.1 101.4 99.2 99 99.7 ...  ..$
X6 : num [1:200] 200 199 200 200 200 ...  ..$ X7 : num [1:200] 199 201 200
199 200 ...  ..$ X8 : num [1:200] 200 201 199 200 201 ...  ..$ X9 : num
[1:200] 199 199 201 201 199 ...  ..$ X10 : num [1:200] 201 201 200 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.3 100.3 98.8 99.4
100.5 ...  ..$ X2 : num [1:200] 98.1 99.1 98.8 98 100.8 ...  ..$ X3 : num
[1:200] 100 101 101 100 101 ...  ..$ X4 : num [1:200] 100.1 99.8 100.6 101.6
101.5 ...  ..$ X5 : num [1:200] 101.7 99.2 99.7 100.3 100.7 ...  ..$ X6 :
num [1:200] 200 201 200 200 200 ...  ..$ X7 : num [1:200] 199 199 199 200
200 ...  ..$ X8 : num [1:200] 200 201 199 200 200 ...  ..$ X9 : num [1:200]
200 201 200 201 200 ...  ..$ X10 : num [1:200] 198 201 200 200 198 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.5 99.7 99.2 100.3
99.7 ...  ..$ X2 : num [1:200] 101.3 100.8 99.6 98.6 101 ...  ..$ X3 : num
[1:200] 100 100 99 101 101 ...  ..$ X4 : num [1:200] 100 101.3 100.8 98.9
98.2 ...  ..$ X5 : num [1:200] 99.3 100.3 100.3 99.8 99.6 ...  ..$ X6 : num
[1:200] 199 199 200 201 200 ...  ..$ X7 : num [1:200] 200 199 200 200 200
...  ..$ X8 : num [1:200] 199 200 200 201 200 ...  ..$ X9 : num [1:200] 198
202 200 200 200 ...  ..$ X10 : num [1:200] 202 201 200 201 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.4 102.3 99 99
101.1 ...  ..$ X2 : num [1:200] 99.8 99.4 99.1 101.4 100 ...  ..$ X3 : num
[1:200] 100.2 99.8 100.2 100.8 98.5 ...  ..$ X4 : num [1:200] 97.7 101 99.4
99.6 99.2 ...  ..$ X5 : num [1:200] 101 100 101 100 101 ...  ..$ X6 : num
[1:200] 201 200 199 199 201 ...  ..$ X7 : num [1:200] 200 199 201 201 201
...  ..$ X8 : num [1:200] 199 201 202 199 201 ...  ..$ X9 : num [1:200] 197
200 199 199 199 ...  ..$ X10 : num [1:200] 201 200 200 199 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.7 98.1 99.3 100.7
99.8 ...  ..$ X2 : num [1:200] 99.7 100.6 99.6 99.7 99.6 ...  ..$ X3 : num
[1:200] 100.8 100.9 100.2 99.6 101.9 ...  ..$ X4 : num [1:200] 100.5 100
99.8 98.6 99.5 ...  ..$ X5 : num [1:200] 101 100.5 100.5 101.9 99.3 ...  ..$
X6 : num [1:200] 201 198 201 198 201 ...  ..$ X7 : num [1:200] 200 200 198
199 199 ...  ..$ X8 : num [1:200] 199 198 200 199 200 ...  ..$ X9 : num
[1:200] 200 201 202 200 201 ...  ..$ X10 : num [1:200] 198 199 199 200 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.1 100.8 99.1 99.7
98.7 ...  ..$ X2 : num [1:200] 100.9 99.7 99.7 100.6 101.2 ...  ..$ X3 : num
[1:200] 99.6 98.2 99.9 99.7 100.5 ...  ..$ X4 : num [1:200] 98.9 99.9 102
100.4 101.7 ...  ..$ X5 : num [1:200] 99.1 99.8 99.5 102.2 100 ...  ..$ X6 :
num [1:200] 200 201 200 200 201 ...  ..$ X7 : num [1:200] 202 201 199 201
198 ...  ..$ X8 : num [1:200] 200 200 199 200 200 ...  ..$ X9 : num [1:200]
201 200 201 199 199 ...  ..$ X10 : num [1:200] 200 202 199 199 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.5 99 99.4 101.5 99
...  ..$ X2 : num [1:200] 97.8 100.4 100.7 99.8 100.4 ...  ..$ X3 : num
[1:200] 99.8 98.2 101 97.9 99.6 ...  ..$ X4 : num [1:200] 100.3 99.3 100.8
99.7 99.2 ...  ..$ X5 : num [1:200] 99.8 99.7 100.6 101.2 100.6 ...  ..$ X6
: num [1:200] 199 199 201 199 199 ...  ..$ X7 : num [1:200] 200 201 200 202
200 ...  ..$ X8 : num [1:200] 199 199 202 200 202 ...  ..$ X9 : num [1:200]
200 199 198 200 199 ...  ..$ X10 : num [1:200] 200 200 200 199 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.4 98.9 100.5 99.5
98.9 ...  ..$ X2 : num [1:200] 99.9 99.9 100 100.7 101 ...  ..$ X3 : num
[1:200] 99.7 99.7 98.4 101.2 100.8 ...  ..$ X4 : num [1:200] 100.9 100.6
100.6 99.1 101.8 ...  ..$ X5 : num [1:200] 99 98.1 101.8 99.5 100.3 ...  ..$
X6 : num [1:200] 200 199 200 198 202 ...  ..$ X7 : num [1:200] 201 199 200
199 198 ...  ..$ X8 : num [1:200] 200 198 201 199 200 ...  ..$ X9 : num
[1:200] 200 200 201 201 199 ...  ..$ X10 : num [1:200] 200 199 200 198 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.6 99 99.4 101.3
101.1 ...  ..$ X2 : num [1:200] 101.1 99.5 99.1 100.7 98.8 ...  ..$ X3 : num
[1:200] 101 101 100 102 101 ...  ..$ X4 : num [1:200] 100.5 98.9 98.5 100.7
99.4 ...  ..$ X5 : num [1:200] 101.4 101 99.2 99.9 101.2 ...  ..$ X6 : num
[1:200] 200 200 199 201 200 ...  ..$ X7 : num [1:200] 201 198 198 201 201
...  ..$ X8 : num [1:200] 200 200 200 200 200 ...  ..$ X9 : num [1:200] 201
201 202 199 200 ...  ..$ X10 : num [1:200] 199 198 201 198 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 101.1 100.4
101.3 101.5 ...  ..$ X2 : num [1:200] 99.4 98.3 102.3 100.7 96.8 ...  ..$ X3
: num [1:200] 101 100 99.5 98.5 100.2 ...  ..$ X4 : num [1:200] 97.8 99.4
100.4 101.4 101.2 ...  ..$ X5 : num [1:200] 100 99.9 100.4 101.2 100.6 ...
..$ X6 : num [1:200] 199 200 201 200 198 ...  ..$ X7 : num [1:200] 200 199
200 200 200 ...  ..$ X8 : num [1:200] 202 199 199 197 200 ...  ..$ X9 : num
[1:200] 199 201 199 201 202 ...  ..$ X10 : num [1:200] 200 199 200 200 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.1 100.3 98.9 99.5
100.1 ...  ..$ X2 : num [1:200] 100.8 99.9 100.5 100.6 98.9 ...  ..$ X3 :
num [1:200] 99.7 100.7 99.9 101 99 ...  ..$ X4 : num [1:200] 101 100 101 103
100 ...  ..$ X5 : num [1:200] 99.2 101.6 100.8 100.4 101.4 ...  ..$ X6 : num
[1:200] 202 200 199 200 200 ...  ..$ X7 : num [1:200] 200 199 199 200 200
...  ..$ X8 : num [1:200] 199 199 200 200 201 ...  ..$ X9 : num [1:200] 200
200 200 199 201 ...  ..$ X10 : num [1:200] 200 200 201 200 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.7 99.7 99.6 100
99.1 ...  ..$ X2 : num [1:200] 99.9 99.5 100.8 99.4 98.5 ...  ..$ X3 : num
[1:200] 99.5 99.3 101.9 100.3 99.2 ...  ..$ X4 : num [1:200] 98.7 101.5
100.9 99.6 100.3 ...  ..$ X5 : num [1:200] 99 99.1 99.9 99.7 99.2 ...  ..$
X6 : num [1:200] 201 199 200 202 200 ...  ..$ X7 : num [1:200] 201 199 200
200 201 ...  ..$ X8 : num [1:200] 201 201 199 200 201 ...  ..$ X9 : num
[1:200] 201 201 201 199 200 ...  ..$ X10 : num [1:200] 202 201 200 200 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 100.5 102.2 99.5
100.4 ...  ..$ X2 : num [1:200] 98.9 100.9 99.4 102.8 100.8 ...  ..$ X3 :
num [1:200] 98 99.5 101.3 101.6 99.4 ...  ..$ X4 : num [1:200] 100.8 100
100.5 100.9 99.5 ...  ..$ X5 : num [1:200] 100 101.4 99.9 98.4 100.5 ...
..$ X6 : num [1:200] 200 200 202 199 200 ...  ..$ X7 : num [1:200] 199 200
200 200 201 ...  ..$ X8 : num [1:200] 200 198 202 201 199 ...  ..$ X9 : num
[1:200] 201 200 200 200 200 ...  ..$ X10 : num [1:200] 200 200 200 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.5 100.2 98.9 101.5
98.7 ...  ..$ X2 : num [1:200] 100.5 100.2 98.6 99.7 102.7 ...  ..$ X3 : num
[1:200] 97.8 99.1 99.5 99.2 99.7 ...  ..$ X4 : num [1:200] 99.2 99.5 99.4 99
99.2 ...  ..$ X5 : num [1:200] 99.2 99.9 101.7 101.2 102.5 ...  ..$ X6 : num
[1:200] 200 200 199 199 199 ...  ..$ X7 : num [1:200] 199 202 200 200 201
...  ..$ X8 : num [1:200] 199 201 199 200 199 ...  ..$ X9 : num [1:200] 201
202 199 201 201 ...  ..$ X10 : num [1:200] 199 200 201 201 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.3 100.8 100.6
100.7 101 ...  ..$ X2 : num [1:200] 97.6 99.5 99.7 100.6 100.2 ...  ..$ X3 :
num [1:200] 99.1 100.7 101.5 98.3 100.7 ...  ..$ X4 : num [1:200] 99.9 99
99.1 100.2 100.2 ...  ..$ X5 : num [1:200] 99.7 99.9 100.4 100.7 99.7 ...
..$ X6 : num [1:200] 198 200 199 200 199 ...  ..$ X7 : num [1:200] 199 200
200 199 201 ...  ..$ X8 : num [1:200] 198 199 197 200 201 ...  ..$ X9 : num
[1:200] 200 201 200 200 200 ...  ..$ X10 : num [1:200] 201 201 201 199 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.1 100.6 100.1
101.2 100.2 ...  ..$ X2 : num [1:200] 100 100 101 101 100 ...  ..$ X3 : num
[1:200] 102.1 99.8 99.7 101.4 100.3 ...  ..$ X4 : num [1:200] 100.2 98.3
99.7 102.1 102 ...  ..$ X5 : num [1:200] 98.4 99.7 99.1 100.6 99.8 ...  ..$
X6 : num [1:200] 201 200 200 201 200 ...  ..$ X7 : num [1:200] 201 201 199
201 202 ...  ..$ X8 : num [1:200] 201 200 198 200 201 ...  ..$ X9 : num
[1:200] 201 199 200 200 201 ...  ..$ X10 : num [1:200] 200 200 199 200 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99 100.1 99.8 99.4
100.2 ...  ..$ X2 : num [1:200] 100.4 98.9 100.8 98.2 100.4 ...  ..$ X3 :
num [1:200] 100.3 99.9 97.7 99.2 100.1 ...  ..$ X4 : num [1:200] 101.3 101.6
99.2 99.3 99.8 ...  ..$ X5 : num [1:200] 101.2 100.2 100 99.4 97.3 ...  ..$
X6 : num [1:200] 202 200 198 201 201 ...  ..$ X7 : num [1:200] 201 201 200
199 201 ...  ..$ X8 : num [1:200] 199 200 199 200 201 ...  ..$ X9 : num
[1:200] 198 199 200 201 201 ...  ..$ X10 : num [1:200] 200 200 201 200 198
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.8 100.7 100.3 99
99.7 ...  ..$ X2 : num [1:200] 101.3 100.7 101.6 100.5 99.5 ...  ..$ X3 :
num [1:200] 98.2 99.3 101.3 100.1 100.6 ...  ..$ X4 : num [1:200] 98.9 99.6
98.8 98.9 100.2 ...  ..$ X5 : num [1:200] 99.3 99.7 99.8 98.7 99.5 ...  ..$
X6 : num [1:200] 199 200 199 200 200 ...  ..$ X7 : num [1:200] 200 201 201
199 198 ...  ..$ X8 : num [1:200] 199 200 201 199 201 ...  ..$ X9 : num
[1:200] 200 199 199 201 202 ...  ..$ X10 : num [1:200] 202 202 201 198 198
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.4 101.6 98.9 101
101.2 ...  ..$ X2 : num [1:200] 99.8 99.1 100.4 98.8 100.8 ...  ..$ X3 : num
[1:200] 100.1 100.3 100.5 99.6 100.4 ...  ..$ X4 : num [1:200] 98.4 100.8
99.8 99.2 98.2 ...  ..$ X5 : num [1:200] 98.8 100.9 100.6 99.8 101.5 ...
..$ X6 : num [1:200] 199 199 200 199 201 ...  ..$ X7 : num [1:200] 198 201
200 198 199 ...  ..$ X8 : num [1:200] 198 199 200 202 200 ...  ..$ X9 : num
[1:200] 199 198 201 200 200 ...  ..$ X10 : num [1:200] 201 201 201 199 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.4 100.3 99.4 100.2
102.2 ...  ..$ X2 : num [1:200] 100.8 99.7 99.7 100.6 101.6 ...  ..$ X3 :
num [1:200] 100.2 100.1 99.5 99.9 100.1 ...  ..$ X4 : num [1:200] 99.7 98.9
101 100.9 100.4 ...  ..$ X5 : num [1:200] 99.4 101.6 99.7 100.3 100 ...  ..$
X6 : num [1:200] 199 200 201 199 200 ...  ..$ X7 : num [1:200] 201 200 199
200 200 ...  ..$ X8 : num [1:200] 201 200 202 201 199 ...  ..$ X9 : num
[1:200] 200 201 200 200 201 ...  ..$ X10 : num [1:200] 200 200 202 202 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 101.1 101.7
100.2 101.1 ...  ..$ X2 : num [1:200] 98.9 101.7 100.1 99.8 101.1 ...  ..$
X3 : num [1:200] 99.9 100.6 99.6 100.4 100.1 ...  ..$ X4 : num [1:200] 99.8
99.8 98.3 98.8 100.6 ...  ..$ X5 : num [1:200] 97.7 99.8 99.1 99.8 98 ...
..$ X6 : num [1:200] 199 199 200 201 201 ...  ..$ X7 : num [1:200] 200 200
200 200 200 ...  ..$ X8 : num [1:200] 199 200 201 199 200 ...  ..$ X9 : num
[1:200] 200 201 200 200 199 ...  ..$ X10 : num [1:200] 202 200 199 201 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.6 99.6 100.3 99.3
101 ...  ..$ X2 : num [1:200] 101.1 100.3 98.6 99.6 99.1 ...  ..$ X3 : num
[1:200] 99.4 98.6 100.3 99.7 100.6 ...  ..$ X4 : num [1:200] 99.3 100 99.3
99.5 99.4 ...  ..$ X5 : num [1:200] 99.4 99.9 98.9 100.6 100.3 ...  ..$ X6 :
num [1:200] 201 200 201 199 200 ...  ..$ X7 : num [1:200] 201 201 199 200
200 ...  ..$ X8 : num [1:200] 199 200 200 203 199 ...  ..$ X9 : num [1:200]
201 200 199 201 200 ...  ..$ X10 : num [1:200] 199 201 200 201 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.1 99.5 100.1
100.6 100.3 ...  ..$ X2 : num [1:200] 99.7 100.8 101 101.4 98.4 ...  ..$ X3
: num [1:200] 100.8 98.7 99.6 100.6 101.4 ...  ..$ X4 : num [1:200] 101.5
99.4 101.2 99.8 99.2 ...  ..$ X5 : num [1:200] 100.5 99.5 97.5 101.6 99.7
...  ..$ X6 : num [1:200] 200 200 201 200 201 ...  ..$ X7 : num [1:200] 199
199 200 200 199 ...  ..$ X8 : num [1:200] 199 202 200 202 200 ...  ..$ X9 :
num [1:200] 201 201 200 201 200 ...  ..$ X10 : num [1:200] 201 199 200 199
200 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 99.7 100.5
99.5 99.9 ...  ..$ X2 : num [1:200] 100.2 99.6 100.5 100.3 100.7 ...  ..$ X3
: num [1:200] 100.1 99.4 98.3 99.2 100.2 ...  ..$ X4 : num [1:200] 98.6
101.1 100.9 100.7 100.8 ...  ..$ X5 : num [1:200] 101 100 100 100 100 ...
..$ X6 : num [1:200] 200 201 199 199 201 ...  ..$ X7 : num [1:200] 201 202
200 199 199 ...  ..$ X8 : num [1:200] 199 199 200 202 201 ...  ..$ X9 : num
[1:200] 199 199 200 200 199 ...  ..$ X10 : num [1:200] 200 199 201 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.1 102.9 99.9 99.5
99.9 ...  ..$ X2 : num [1:200] 101.5 101.8 99.5 99.2 99 ...  ..$ X3 : num
[1:200] 99.8 98.6 99.7 100 99.6 ...  ..$ X4 : num [1:200] 101.8 101.1 99.2
100 98.5 ...  ..$ X5 : num [1:200] 100.5 101.3 101 100.7 97.6 ...  ..$ X6 :
num [1:200] 198 200 200 199 200 ...  ..$ X7 : num [1:200] 201 200 201 199
200 ...  ..$ X8 : num [1:200] 198 201 198 200 200 ...  ..$ X9 : num [1:200]
200 199 202 201 200 ...  ..$ X10 : num [1:200] 200 199 198 200 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.7 100.2 99.5
101.7 100.4 ...  ..$ X2 : num [1:200] 100.3 100.4 101.5 101 99.7 ...  ..$ X3
: num [1:200] 100.4 100.1 98.2 101.9 100.5 ...  ..$ X4 : num [1:200] 100.6
102.2 101.4 99.6 98.6 ...  ..$ X5 : num [1:200] 101 100.6 97.5 99.6 100.9
...  ..$ X6 : num [1:200] 201 200 200 199 201 ...  ..$ X7 : num [1:200] 199
200 200 201 200 ...  ..$ X8 : num [1:200] 199 200 199 199 200 ...  ..$ X9 :
num [1:200] 200 199 201 200 199 ...  ..$ X10 : num [1:200] 199 200 201 199
201 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101 98.7 99.4
101.7 101.3 ...  ..$ X2 : num [1:200] 99.3 100.7 98.5 100.9 101.7 ...  ..$
X3 : num [1:200] 98.9 97.9 100.5 101 99.6 ...  ..$ X4 : num [1:200] 98.9
100.4 99 99.9 100.8 ...  ..$ X5 : num [1:200] 100.7 101.2 101.1 98.6 99.6
...  ..$ X6 : num [1:200] 202 198 199 199 200 ...  ..$ X7 : num [1:200] 199
201 200 199 200 ...  ..$ X8 : num [1:200] 200 200 200 199 200 ...  ..$ X9 :
num [1:200] 201 199 201 199 200 ...  ..$ X10 : num [1:200] 201 199 199 200
199 ...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: ..
..$ Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition :
Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10]
1 2 3 4 5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs:
int [1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.4 100.1 100.5
99.9 98.9 ...  ..$ X2 : num [1:200] 100 99.4 100.3 99.4 98.3 ...  ..$ X3 :
num [1:200] 100.1 101.1 99.7 99.9 98.9 ...  ..$ X4 : num [1:200] 100.2 100.9
99.8 100 99.6 ...  ..$ X5 : num [1:200] 98.5 99.5 99.5 98.9 99.9 ...  ..$ X6
: num [1:200] 197 200 199 201 198 ...  ..$ X7 : num [1:200] 199 200 200 198
200 ...  ..$ X8 : num [1:200] 200 200 198 200 201 ...  ..$ X9 : num [1:200]
200 200 200 199 200 ...  ..$ X10 : num [1:200] 200 200 201 201 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.9 99.4 99.8 99.2
99.9 ...  ..$ X2 : num [1:200] 101.9 98.2 99.6 99.7 101 ...  ..$ X3 : num
[1:200] 99.5 100.4 101.1 99 100 ...  ..$ X4 : num [1:200] 100.6 99.5 100.5
99.9 98.6 ...  ..$ X5 : num [1:200] 101.2 100 99.1 101.1 100.6 ...  ..$ X6 :
num [1:200] 201 197 199 199 199 ...  ..$ X7 : num [1:200] 199 200 201 201
200 ...  ..$ X8 : num [1:200] 200 200 199 199 200 ...  ..$ X9 : num [1:200]
202 199 200 199 201 ...  ..$ X10 : num [1:200] 202 200 200 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 97.9 100.4 100.7 99.4
99.1 ...  ..$ X2 : num [1:200] 100 99 99 99.1 101.7 ...  ..$ X3 : num
[1:200] 100.8 99.7 100.4 99 98.5 ...  ..$ X4 : num [1:200] 101 99.9 99.2
99.4 101.5 ...  ..$ X5 : num [1:200] 101 101 98.2 100.3 101.7 ...  ..$ X6 :
num [1:200] 200 200 200 199 200 ...  ..$ X7 : num [1:200] 200 199 201 199
202 ...  ..$ X8 : num [1:200] 200 200 200 201 200 ...  ..$ X9 : num [1:200]
200 201 200 202 198 ...  ..$ X10 : num [1:200] 200 200 201 200 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99 100 101.8 100.2
99.3 ...  ..$ X2 : num [1:200] 99.5 99.5 98.1 100.5 101.2 ...  ..$ X3 : num
[1:200] 98.9 97.6 98.5 100.8 99.8 ...  ..$ X4 : num [1:200] 101.2 100.6 99.8
101.6 99.8 ...  ..$ X5 : num [1:200] 100 102 101 102 101 ...  ..$ X6 : num
[1:200] 203 201 200 202 200 ...  ..$ X7 : num [1:200] 199 201 201 200 199
...  ..$ X8 : num [1:200] 200 199 199 200 203 ...  ..$ X9 : num [1:200] 201
198 200 201 199 ...  ..$ X10 : num [1:200] 201 200 200 202 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.1 100 100.5 99.7
99.9 ...  ..$ X2 : num [1:200] 99.8 98.5 101.2 98.6 100.8 ...  ..$ X3 : num
[1:200] 101.3 100.5 101.1 99.8 100.3 ...  ..$ X4 : num [1:200] 100.7 100.4
100.4 100.2 98.9 ...  ..$ X5 : num [1:200] 100.3 99.7 101 100.9 99.2 ...
..$ X6 : num [1:200] 199 201 199 201 199 ...  ..$ X7 : num [1:200] 200 200
200 201 200 ...  ..$ X8 : num [1:200] 200 200 200 201 200 ...  ..$ X9 : num
[1:200] 198 198 201 201 202 ...  ..$ X10 : num [1:200] 201 200 199 199 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.5 99.1 100 99.7
98.3 ...  ..$ X2 : num [1:200] 99.6 99.4 99.8 99.3 101.3 ...  ..$ X3 : num
[1:200] 98.8 97.6 98.8 100.1 99.8 ...  ..$ X4 : num [1:200] 100.9 99.2 99.7
98.9 100.4 ...  ..$ X5 : num [1:200] 99.6 99.7 101 102.1 100.1 ...  ..$ X6 :
num [1:200] 200 200 201 200 201 ...  ..$ X7 : num [1:200] 200 198 200 200
200 ...  ..$ X8 : num [1:200] 201 200 198 200 200 ...  ..$ X9 : num [1:200]
200 200 198 201 200 ...  ..$ X10 : num [1:200] 200 200 201 202 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.7 100.6 98.4 99
101.1 ...  ..$ X2 : num [1:200] 100.2 100.9 100.4 99.6 98.6 ...  ..$ X3 :
num [1:200] 99.4 99.3 101.4 99.2 100.3 ...  ..$ X4 : num [1:200] 99.6 101.3
100 99.6 100.5 ...  ..$ X5 : num [1:200] 100 101 101 100 101 ...  ..$ X6 :
num [1:200] 199 200 200 200 199 ...  ..$ X7 : num [1:200] 199 201 199 200
199 ...  ..$ X8 : num [1:200] 200 200 202 198 199 ...  ..$ X9 : num [1:200]
201 200 200 200 201 ...  ..$ X10 : num [1:200] 200 200 200 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.4 99.7 99.9 99.6
100.8 ...  ..$ X2 : num [1:200] 100.9 99.3 101 98.7 100.9 ...  ..$ X3 : num
[1:200] 100.4 98.5 99.7 98.5 98.5 ...  ..$ X4 : num [1:200] 99.6 100.2 101
99.3 99.9 ...  ..$ X5 : num [1:200] 100.8 101.1 99.6 99.3 98.8 ...  ..$ X6 :
num [1:200] 201 199 200 200 202 ...  ..$ X7 : num [1:200] 200 200 201 200
201 ...  ..$ X8 : num [1:200] 199 199 200 201 200 ...  ..$ X9 : num [1:200]
200 200 201 201 200 ...  ..$ X10 : num [1:200] 202 200 200 199 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.1 99.2 100.5
100.1 100.1 ...  ..$ X2 : num [1:200] 99.6 99.2 99.6 101.3 100.6 ...  ..$ X3
: num [1:200] 97.5 99 99 99 100.4 ...  ..$ X4 : num [1:200] 98.8 98.5 100
102.3 99.4 ...  ..$ X5 : num [1:200] 99 101 102 100 100 ...  ..$ X6 : num
[1:200] 203 201 200 201 199 ...  ..$ X7 : num [1:200] 201 198 200 201 200
...  ..$ X8 : num [1:200] 201 200 202 200 200 ...  ..$ X9 : num [1:200] 201
200 200 200 200 ...  ..$ X10 : num [1:200] 199 201 199 201 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.8 100.2 100.5 99.9
99.9 ...  ..$ X2 : num [1:200] 99.4 99.8 100.9 101.9 99.3 ...  ..$ X3 : num
[1:200] 99 101.1 102.2 99.7 99.6 ...  ..$ X4 : num [1:200] 99.5 99.7 100.6
98.3 99.5 ...  ..$ X5 : num [1:200] 99.4 100.8 100.6 99.9 100.3 ...  ..$ X6
: num [1:200] 200 201 200 200 200 ...  ..$ X7 : num [1:200] 201 202 200 199
201 ...  ..$ X8 : num [1:200] 199 201 198 200 201 ...  ..$ X9 : num [1:200]
199 200 201 200 200 ...  ..$ X10 : num [1:200] 199 198 201 202 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.9 100.7 99.7 100.2
99.4 ...  ..$ X2 : num [1:200] 100.8 99.1 100.3 99.9 101.6 ...  ..$ X3 : num
[1:200] 101.7 99.9 101.5 101.1 99.6 ...  ..$ X4 : num [1:200] 99.4 99.8 101
101.4 102.5 ...  ..$ X5 : num [1:200] 99.2 101.6 101.7 101 101 ...  ..$ X6 :
num [1:200] 199 201 200 200 201 ...  ..$ X7 : num [1:200] 200 201 200 200
200 ...  ..$ X8 : num [1:200] 200 203 200 200 200 ...  ..$ X9 : num [1:200]
199 201 201 198 200 ...  ..$ X10 : num [1:200] 201 203 201 200 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.7 100.4 99.8 99.6
99.8 ...  ..$ X2 : num [1:200] 100.4 100.7 99.7 100 100.1 ...  ..$ X3 : num
[1:200] 99.4 99.6 99.4 99.9 101.5 ...  ..$ X4 : num [1:200] 100.2 99.9 100.6
100 100.2 ...  ..$ X5 : num [1:200] 99.5 100.6 100.5 100.3 98.9 ...  ..$ X6
: num [1:200] 198 200 201 200 201 ...  ..$ X7 : num [1:200] 200 200 200 199
201 ...  ..$ X8 : num [1:200] 201 199 201 199 200 ...  ..$ X9 : num [1:200]
197 199 200 200 200 ...  ..$ X10 : num [1:200] 198 199 200 200 198 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.3 98.9 100.3 99.6
99.7 ...  ..$ X2 : num [1:200] 101.7 100.2 99.2 101.6 98.8 ...  ..$ X3 : num
[1:200] 100.5 99.8 100.4 98.3 98.5 ...  ..$ X4 : num [1:200] 100.2 99.4 99.6
101.2 99.3 ...  ..$ X5 : num [1:200] 99.3 99.3 99.6 101.5 99.5 ...  ..$ X6 :
num [1:200] 200 198 201 199 201 ...  ..$ X7 : num [1:200] 198 201 201 200
198 ...  ..$ X8 : num [1:200] 199 200 199 201 199 ...  ..$ X9 : num [1:200]
199 200 200 201 199 ...  ..$ X10 : num [1:200] 200 200 197 200 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.5 100.8 99.6
100.5 99.6 ...  ..$ X2 : num [1:200] 99.6 101.3 100.4 100.6 98.5 ...  ..$ X3
: num [1:200] 100.3 101 98.6 100.9 99.8 ...  ..$ X4 : num [1:200] 98.9 98.8
99.8 100.5 101.5 ...  ..$ X5 : num [1:200] 99.6 101.3 102 101.7 101.2 ...
..$ X6 : num [1:200] 200 200 199 199 201 ...  ..$ X7 : num [1:200] 200 202
200 200 198 ...  ..$ X8 : num [1:200] 200 202 200 200 202 ...  ..$ X9 : num
[1:200] 199 201 200 201 201 ...  ..$ X10 : num [1:200] 200 201 201 200 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.9 100.2 99.5 100
99.5 ...  ..$ X2 : num [1:200] 100.5 100 103.2 99.5 102 ...  ..$ X3 : num
[1:200] 99.7 100 99.5 100.1 98.7 ...  ..$ X4 : num [1:200] 101.8 99.1 99.6
100.8 100.9 ...  ..$ X5 : num [1:200] 99.6 100.7 99.3 98.9 101.2 ...  ..$ X6
: num [1:200] 200 200 201 199 200 ...  ..$ X7 : num [1:200] 200 200 199 200
200 ...  ..$ X8 : num [1:200] 200 200 198 199 199 ...  ..$ X9 : num [1:200]
201 200 199 201 200 ...  ..$ X10 : num [1:200] 202 200 201 201 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.4 99.7 99.4 98.8
99.4 ...  ..$ X2 : num [1:200] 100.2 100 99 99.6 100.9 ...  ..$ X3 : num
[1:200] 100.6 98.1 99.1 100.2 99.6 ...  ..$ X4 : num [1:200] 98.7 100.1
100.8 99.9 99.5 ...  ..$ X5 : num [1:200] 98.9 98.1 100.1 101.6 98.7 ...
..$ X6 : num [1:200] 201 200 200 198 201 ...  ..$ X7 : num [1:200] 199 199
201 201 201 ...  ..$ X8 : num [1:200] 199 200 201 199 201 ...  ..$ X9 : num
[1:200] 200 198 200 201 200 ...  ..$ X10 : num [1:200] 198 200 201 199 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.9 99.4 99.4 101.3
101.2 ...  ..$ X2 : num [1:200] 100.8 99.1 99.8 100.1 101.2 ...  ..$ X3 :
num [1:200] 100.8 100 98.5 100 100 ...  ..$ X4 : num [1:200] 98.5 100.4
101.1 99.4 101.3 ...  ..$ X5 : num [1:200] 100.5 99.8 99.6 100.5 98.2 ...
..$ X6 : num [1:200] 200 199 201 199 200 ...  ..$ X7 : num [1:200] 201 201
199 200 201 ...  ..$ X8 : num [1:200] 200 201 199 201 201 ...  ..$ X9 : num
[1:200] 203 199 200 199 201 ...  ..$ X10 : num [1:200] 200 202 200 200 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.3 101.2 99.9 98.8
101.2 ...  ..$ X2 : num [1:200] 100.6 102 100.7 99.6 99.5 ...  ..$ X3 : num
[1:200] 97.6 100.6 99.5 99.4 101.4 ...  ..$ X4 : num [1:200] 101.5 102.1
98.9 99.5 101.2 ...  ..$ X5 : num [1:200] 101 100 102 100 100 ...  ..$ X6 :
num [1:200] 199 199 202 200 199 ...  ..$ X7 : num [1:200] 201 201 199 200
200 ...  ..$ X8 : num [1:200] 202 201 200 200 200 ...  ..$ X9 : num [1:200]
201 200 201 199 201 ...  ..$ X10 : num [1:200] 200 200 201 200 202 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 98.4 100.1 100.6 99.7
99.4 ...  ..$ X2 : num [1:200] 100.4 100.6 99.4 101 98.8 ...  ..$ X3 : num
[1:200] 101 99.7 100.5 100.9 100.6 ...  ..$ X4 : num [1:200] 100.2 101.6 100
100.3 99.7 ...  ..$ X5 : num [1:200] 98.7 100.8 100.8 101.2 99.8 ...  ..$ X6
: num [1:200] 202 197 199 200 199 ...  ..$ X7 : num [1:200] 200 199 200 200
200 ...  ..$ X8 : num [1:200] 200 199 198 201 199 ...  ..$ X9 : num [1:200]
199 199 199 198 202 ...  ..$ X10 : num [1:200] 200 199 201 199 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.7 100.8 100.7
97.7 100.5 ...  ..$ X2 : num [1:200] 99.8 100.4 101.4 98.8 100.4 ...  ..$ X3
: num [1:200] 99.9 99.7 100.4 100.8 100 ...  ..$ X4 : num [1:200] 98.9 100.2
100.7 99.8 100.8 ...  ..$ X5 : num [1:200] 100 100 101 100 101 ...  ..$ X6 :
num [1:200] 200 201 202 199 200 ...  ..$ X7 : num [1:200] 200 202 202 198
201 ...  ..$ X8 : num [1:200] 200 201 200 198 199 ...  ..$ X9 : num [1:200]
197 200 200 201 199 ...  ..$ X10 : num [1:200] 199 198 199 198 200 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.5 100.3 98.4
100.1 100.8 ...  ..$ X2 : num [1:200] 100.6 100.3 99.3 99.4 98.6 ...  ..$ X3
: num [1:200] 98.9 99.8 98.9 100.1 98.6 ...  ..$ X4 : num [1:200] 100.2
100.2 101 99.9 100.6 ...  ..$ X5 : num [1:200] 98.2 99.6 99.6 99.7 100 ...
..$ X6 : num [1:200] 201 199 200 200 200 ...  ..$ X7 : num [1:200] 200 201
200 200 200 ...  ..$ X8 : num [1:200] 200 200 202 200 201 ...  ..$ X9 : num
[1:200] 200 202 199 200 201 ...  ..$ X10 : num [1:200] 202 201 201 201 199
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 99.6 99.4 100.5 100.1
100.1 ...  ..$ X2 : num [1:200] 100.2 101 98.9 99.9 101.1 ...  ..$ X3 : num
[1:200] 100 100 101 102 100 ...  ..$ X4 : num [1:200] 98.7 99.8 98.8 100 102
...  ..$ X5 : num [1:200] 98.4 100.3 100.9 100.7 99.3 ...  ..$ X6 : num
[1:200] 199 201 200 200 200 ...  ..$ X7 : num [1:200] 200 199 201 202 201
...  ..$ X8 : num [1:200] 198 199 200 199 201 ...  ..$ X9 : num [1:200] 199
200 202 199 200 ...  ..$ X10 : num [1:200] 200 201 199 200 198 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.5 100.5 99 99.3
101.8 ...  ..$ X2 : num [1:200] 100 100 101 100 101 ...  ..$ X3 : num
[1:200] 99 101.6 101 99.7 100.9 ...  ..$ X4 : num [1:200] 100.4 99.8 100.1
96.9 100.1 ...  ..$ X5 : num [1:200] 99.7 99.8 100 99.8 99.2 ...  ..$ X6 :
num [1:200] 201 201 199 201 200 ...  ..$ X7 : num [1:200] 199 201 201 200
200 ...  ..$ X8 : num [1:200] 200 199 199 199 200 ...  ..$ X9 : num [1:200]
202 200 199 200 199 ...  ..$ X10 : num [1:200] 198 198 198 198 199 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101.2 99.8 99.9 99.9
100.5 ...  ..$ X2 : num [1:200] 101.1 99.6 99.4 100.7 100.7 ...  ..$ X3 :
num [1:200] 101.1 100.2 100.7 98.3 99.3 ...  ..$ X4 : num [1:200] 100.9
100.7 100.3 100 99.5 ...  ..$ X5 : num [1:200] 101.9 99.1 97.1 100.4 99 ...
..$ X6 : num [1:200] 200 201 200 201 201 ...  ..$ X7 : num [1:200] 200 201
200 198 201 ...  ..$ X8 : num [1:200] 201 199 200 199 200 ...  ..$ X9 : num
[1:200] 201 200 199 201 200 ...  ..$ X10 : num [1:200] 199 200 201 201 200
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.6 98.6 101.3 99.6
101.2 ...  ..$ X2 : num [1:200] 99.8 98.6 98.5 100.2 100.1 ...  ..$ X3 : num
[1:200] 99.4 100.8 100.7 101.4 98.9 ...  ..$ X4 : num [1:200] 100.8 99.4
100.7 98.6 99.7 ...  ..$ X5 : num [1:200] 100.4 101.1 99.5 100.8 100.2 ...
..$ X6 : num [1:200] 201 199 201 198 200 ...  ..$ X7 : num [1:200] 201 199
200 200 198 ...  ..$ X8 : num [1:200] 199 199 199 199 201 ...  ..$ X9 : num
[1:200] 201 199 199 202 200 ...  ..$ X10 : num [1:200] 201 200 200 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.2 100.4 99.8
100.4 99.6 ...  ..$ X2 : num [1:200] 98 100.1 99.5 99.6 99.7 ...  ..$ X3 :
num [1:200] 100.9 100.1 99.7 100.2 98.8 ...  ..$ X4 : num [1:200] 97.5 100.3
100.6 99.7 100.5 ...  ..$ X5 : num [1:200] 101 101 101 100 102 ...  ..$ X6 :
num [1:200] 202 199 199 200 201 ...  ..$ X7 : num [1:200] 201 203 200 199
199 ...  ..$ X8 : num [1:200] 199 199 199 201 199 ...  ..$ X9 : num [1:200]
200 199 200 199 201 ...  ..$ X10 : num [1:200] 199 199 201 199 201 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101 100.2 99.5 99.9
101.3 ...  ..$ X2 : num [1:200] 100.2 100.5 99.4 100.2 100.6 ...  ..$ X3 :
num [1:200] 99.8 100.1 99.4 100.4 100.1 ...  ..$ X4 : num [1:200] 99.1 99
99.1 100.9 100.4 ...  ..$ X5 : num [1:200] 99 99.9 98.9 100.3 99.4 ...  ..$
X6 : num [1:200] 200 201 198 200 200 ...  ..$ X7 : num [1:200] 199 200 201
199 201 ...  ..$ X8 : num [1:200] 200 199 201 200 200 ...  ..$ X9 : num
[1:200] 201 201 200 201 199 ...  ..$ X10 : num [1:200] 200 201 199 201 201
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 101 99.8 98.7 99.8
99.2 ...  ..$ X2 : num [1:200] 100.4 100.3 99.8 100.3 99.5 ...  ..$ X3 : num
[1:200] 98.9 100.4 98.6 99.5 100.2 ...  ..$ X4 : num [1:200] 100.4 100.7 100
99.8 100.2 ...  ..$ X5 : num [1:200] 100.2 99.9 99.7 99.3 100.6 ...  ..$ X6
: num [1:200] 200 200 201 200 198 ...  ..$ X7 : num [1:200] 201 201 197 199
199 ...  ..$ X8 : num [1:200] 200 201 199 200 199 ...  ..$ X9 : num [1:200]
199 198 200 200 199 ...  ..$ X10 : num [1:200] 200 201 199 200 202 ...  ..-
attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.8 100.2 101.2
99.4 98.9 ...  ..$ X2 : num [1:200] 99.4 99.5 98.8 100.5 100.4 ...  ..$ X3 :
num [1:200] 98.9 99.7 99.7 99.2 101 ...  ..$ X4 : num [1:200] 100.6 98.9
99.6 100.3 101 ...  ..$ X5 : num [1:200] 102.2 99.6 99.7 100.2 101 ...  ..$
X6 : num [1:200] 200 200 200 201 199 ...  ..$ X7 : num [1:200] 199 200 199
200 200 ...  ..$ X8 : num [1:200] 199 201 200 200 200 ...  ..$ X9 : num
[1:200] 202 198 199 200 200 ...  ..$ X10 : num [1:200] 201 200 200 201 202
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 $ :'data.frame': 200 obs. of 11 variables: ..$ id.obs: int
[1:200] 1 2 3 4 5 6 7 8 9 10 ...  ..$ X1 : num [1:200] 100.3 98.1 100.4
100.6 99.7 ...  ..$ X2 : num [1:200] 100.8 99.3 100 97.5 99.8 ...  ..$ X3 :
num [1:200] 101.1 97.9 99.4 102 101.1 ...  ..$ X4 : num [1:200] 98.2 98.9
100.9 101.8 100.8 ...  ..$ X5 : num [1:200] 100.4 98.3 102 98.5 98.8 ...
..$ X6 : num [1:200] 200 200 202 200 199 ...  ..$ X7 : num [1:200] 199 201
201 198 200 ...  ..$ X8 : num [1:200] 201 200 200 202 200 ...  ..$ X9 : num
[1:200] 202 199 201 199 200 ...  ..$ X10 : num [1:200] 199 198 200 202 198
...  ..- attr(*, "metadata")='data.frame': 10 obs. of 3 variables: .. ..$
Sample.name: chr [1:10] "X1" "X2" "X3" "X4" ...  .. ..$ Condition : Factor
w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2 .. ..$ Bio.Rep : int [1:10] 1 2 3 4
5 6 7 8 9 10 [list output truncated]
}
\source{
We simulated the data.
}
\description{
This list of 100 datasets was simulated using the default 
values of the options of the protdatasim function and the set.seed value set 
to 4619.
}
\examples{

data(norm.200.m100.sd1.vs.m200.sd1.list)
str(norm.200.m100.sd1.vs.m200.sd1.list)

}
\references{
arxiv preprint.
}
\author{
M. Chion, Ch. Carapito and F. Bertrand.
}
\keyword{datasets}
