% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_matrix.R
\name{proj_matrix}
\alias{proj_matrix}
\title{Title}
\usage{
proj_matrix(VarRubin.matrix, metadata)
}
\arguments{
\item{VarRubin.matrix}{}

\item{metadata}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
