% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin1_all.R
\name{rubin1.all}
\alias{rubin1.all}
\title{Title}
\usage{
rubin1.all(data, metacond, funcmean = meanImp_emmeans, is.parallel = FALSE)
}
\arguments{
\item{data}{}

\item{metacond}{}

\item{funcmean}{}

\item{is.parallel}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
