% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin1_one.R
\name{rubin1.one}
\alias{rubin1.one}
\title{Title}
\usage{
rubin1.one(peptide, data, funcmean = meanImp_emmeans, metacond)
}
\arguments{
\item{peptide}{}

\item{data}{}

\item{funcmean}{}

\item{metacond}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
