% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin2bt_all.R
\name{rubin2bt.all}
\alias{rubin2bt.all}
\title{Title}
\usage{
rubin2bt.all(
  data,
  funcmean = meanImp_emmeans,
  metacond,
  is.parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{funcmean}{}

\item{metacond}{}

\item{is.parallel}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
