% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin2bt_one.R
\name{rubin2bt.one}
\alias{rubin2bt.one}
\title{Title}
\usage{
rubin2bt.one(peptide, data, funcmean, metacond)
}
\arguments{
\item{peptide}{}

\item{data}{}

\item{funcmean}{}

\item{metacond}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
