% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin2wt_all.R
\name{rubin2wt.all}
\alias{rubin2wt.all}
\title{Title}
\usage{
rubin2wt.all(
  data,
  funcvar = within_variance_comp_emmeans,
  metacond,
  is.parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{}

\item{funcvar}{}

\item{metacond}{}

\item{is.parallel}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
