% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin2wt_one.R
\name{rubin2wt.one}
\alias{rubin2wt.one}
\title{Title}
\usage{
rubin2wt.one(peptide, data, funcvar, metacond)
}
\arguments{
\item{peptide}{}

\item{data}{}

\item{funcvar}{}

\item{metacond}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
