% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.conditions.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design is valid}
\usage{
check.conditions(conds)
}
\arguments{
\item{conds}{A vector}
}
\value{
A list
}
\description{
This function checks the validity of the conditions.

This function was included from the \code{check.conditions} 
function in the \code{DAPAR} package, since \code{DAPAR} was to be removed from 
Bioconductor >= 3.15.
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata')
check.conditions(Biobase::pData(Exp1_R25_pept)$Condition)
}

}
\author{
Samuel Wieczorek as the author of 
\code{check.conditions} in the \code{DAPAR} package.
}
