% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLimmaResult.R
\name{formatLimmaResult}
\alias{formatLimmaResult}
\title{Title}
\usage{
formatLimmaResult(fit, conds, contrast)
}
\arguments{
\item{fit}{}

\item{conds}{}

\item{contrast}{}
}
\value{

}
\description{
It is not exported by DAPAR and has to be reproduced here.
}
\examples{
1+1
}
