% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multi.impute}
\alias{multi.impute}
\title{Title}
\usage{
multi.impute(data, conditions, nb.imp = NULL, method, parallel = F)
}
\arguments{
\item{data}{}

\item{conditions}{}

\item{nb.imp}{}

\item{method}{}

\item{parallel}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
