% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{norm.200.m100.sd1.vs.m200.sd1.list}
\alias{norm.200.m100.sd1.vs.m200.sd1.list}
\title{A list of simulated datasets.}
\format{
The format is: List of 100 data.frames.
\describe{ 
\item{data.frame}{200 obs. of 11 variables
\describe{ 
\item{id.obs}{int [1:200] 1 2 3 4 5 6 7 8 9 10 ...}  
\item{X1}{num [1:200] 99.6 99.9 100.2 99.8 100.4 ...}  
\item{X2}{num [1:200] 97.4 101.3 100.3 100.2 101.7 ...}  
\item{X3}{num [1:200] 100.3 100.9 99.1 101.2 100.6 ...}
\item{X4}{num [1:200] 99.4 99.2 98.5 99.1 99.5 ...}  
\item{X5}{num [1:200] 98.5 99.7 100 100.2 100.7 ...}  
\item{X6}{num [1:200] 200 199 199 200 199 ...}  
\item{X7}{num [1:200] 200 200 202 199 199 ...}  
\item{X8}{num [1:200] 202 199 200 199 201 ...}  
\item{X9}{num [1:200] 200 200 199 201 200 ...}  
\item{X10}{num [1:200] 200 198 200 201 199 ...}
}}
\item{attr(*, "metadata")}{'data.frame': 10 obs. of 3 variables:
\describe{ 
\item{Sample.name}{chr [1:10] "X1" "X2" "X3" "X4" ...}
\item{Condition}{Factor w/ 2 levels "A","B": 1 1 1 1 1 2 2 2 2 2} 
\item{Bio.Rep}{int [1:10] 1 2 3 4 5 6 7 8 9 10}
}}
\item{...}{...}
}
}
\source{
We simulated the data.
}
\description{
This list of 100 datasets was simulated using the default 
values of the options of the protdatasim function and the set.seed value set 
to 4619.
}
\examples{

data(norm.200.m100.sd1.vs.m200.sd1.list)
str(norm.200.m100.sd1.vs.m200.sd1.list)

}
\references{
M. Chion, Ch. Carapito and F. Bertrand (2021). \emph{Accounting for multiple imputation-induced variability for differential analysis in mass spectrometry-based label-free quantitative proteomics}.  \doi{doi:10.1371/journal.pcbi.1010420}.
}
\author{
M. Chion, Ch. Carapito and F. Bertrand.
}
\keyword{datasets}
