% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubin2_all.R
\name{rubin2.all}
\alias{rubin2.all}
\title{Title}
\usage{
rubin2.all(
  data,
  metacond,
  funcmean = meanImp_emmeans,
  funcvar = within_variance_comp_emmeans,
  is.parallel = FALSE
)
}
\arguments{
\item{data}{}

\item{metacond}{}

\item{funcmean}{}

\item{funcvar}{}

\item{is.parallel}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
