% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_variance_comp_emmeans.R
\name{within_variance_comp_emmeans}
\alias{within_variance_comp_emmeans}
\title{Title}
\usage{
within_variance_comp_emmeans(ind, peptide, data, metacond)
}
\arguments{
\item{ind}{}

\item{peptide}{}

\item{data}{}

\item{metacond}{}
}
\value{

}
\description{
Title
}
\examples{
1+1
}
